refactor(db): squash shareable team migrations
