--- observal-server/alembic/versions/025_team_invites.py
@@ -1,61 +1,0 @@
-# SPDX-FileCopyrightText: 2026 Lokesh Selvam <lokeshselvam7025@gmail.com>
-# SPDX-FileCopyrightText: 2026 Hari Srinivasan <harisrini21@gmail.com>
-# SPDX-License-Identifier: Apache-2.0
-
-"""Add private-team invitation links.
-
-Revision ID: 025_invites
-Revises: 024_team_membership_requests
-"""
-
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-from alembic import op
-
-revision = "025_invites"
-down_revision = "024_team_membership_requests"
-branch_labels = None
-depends_on = None
-
-
-def _has_table(name: str) -> bool:
-    return name in sa.inspect(op.get_bind()).get_table_names()
-
-
-def upgrade() -> None:
-    if _has_table("team_invites"):
-        return
-    op.create_table(
-        "team_invites",
-        sa.Column("id", postgresql.UUID(as_uuid=True), primary_key=True),
-        sa.Column("token_hash", sa.String(length=64), nullable=False),
-        sa.Column(
-            "team_id",
-            postgresql.UUID(as_uuid=True),
-            sa.ForeignKey("teams.id", ondelete="CASCADE"),
-            nullable=False,
-        ),
-        sa.Column(
-            "invited_by",
-            postgresql.UUID(as_uuid=True),
-            sa.ForeignKey("users.id", ondelete="SET NULL"),
-            nullable=True,
-        ),
-        sa.Column("max_uses", sa.Integer(), nullable=True),
-        sa.Column("use_count", sa.Integer(), nullable=False, server_default="0"),
-        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("revoked_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
-        sa.UniqueConstraint("token_hash", name="uq_team_invites_token_hash"),
-    )
-    op.create_index("ix_team_invites_team_id", "team_invites", ["team_id"])
-    op.create_index("ix_team_invites_invited_by", "team_invites", ["invited_by"])
-
-
-def downgrade() -> None:
-    if not _has_table("team_invites"):
-        return
-    op.drop_index("ix_team_invites_invited_by", table_name="team_invites")
-    op.drop_index("ix_team_invites_team_id", table_name="team_invites")
-    op.drop_table("team_invites")
